Running a Git Command

Running a Git Command

git(
  ...,
  input = NULL,
  wd = NULL,
  env = NULL,
  timeout = Inf,
  want_process = FALSE
)

Arguments

...
the arguments pass to git command, supports the big bang operator !!!

input
text pass to stdin

wd
working directory

env
additional environment variable pass to git

timeout
throw an error after this amount of time in second

want_process
return the underlying process object instead of the output

Value

The stdout of the process in a scalar character. It may contain a trailing newline. Use trimws() to ensure the trailing newline is trimmed.

Details

If git() fails to locate the git binary automatically, the path could be specify with options(git_path = "/path/to/git").

Examples

# \donttest{
cwd <- getwd()
td <- file.path(tempdir(), "git-examples")
dir.create(td, showWarnings = FALSE)
setwd(td)

git("init")
#> Initialized empty Git repository in /private/var/folders/w9/sklv856j2ks_783rft8mw56c00syhy/T/RtmpB0QdZ6/git-examples/.git/
git("config", "--local", "user.email", "you@example.com")
git("config", "--local", "user.name", "Your Name")

cat("hello\n", file = "hello.txt")

git("status")
#> On branch main
#> 
#> No commits yet
#> 
#> Untracked files:
#>   (use "git add <file>..." to include in what will be committed)
#> 	hello.txt
#> 
#> nothing added to commit but untracked files present (use "git add" to track)

git("add", "hello.txt")

git("commit", "-m", "initial message")
#> [main (root-commit) eec7cfd] initial message
#>  1 file changed, 1 insertion(+)
#>  create mode 100644 hello.txt

git("log")
#> commit eec7cfd7fd1b9d14e3390bcf40565303c66702ea
#> Author: Your Name <you@example.com>
#> Date:   Sat Jun 5 13:12:41 2021 -0700
#> 
#>     initial message

cat("world\n", file = "world.txt")

git("add", "world.txt")

git("commit", "-F", "-", input = "second message")
#> [main f7b0be6] second message
#>  1 file changed, 1 insertion(+)
#>  create mode 100644 world.txt

(sha <- git("rev-parse", "HEAD"))
#> f7b0be6c9e5f39ec4da264bbb70948f1529fc646

git("show", sha)
#> commit f7b0be6c9e5f39ec4da264bbb70948f1529fc646
#> Author: Your Name <you@example.com>
#> Date:   Sat Jun 5 13:12:41 2021 -0700
#> 
#>     second message
#> 
#> diff --git a/world.txt b/world.txt
#> new file mode 100644
#> index 0000000..cc628cc
#> --- /dev/null
#> +++ b/world.txt
#> @@ -0,0 +1 @@
#> +world

setwd(cwd)
# }